Moved ATR files to render/atr/
